LIB-820 Run 2.x/master test suite on x64 simulator

+++ Airship/Common/UAAnalytics+Internal.h
@@ -28,20 +28,20 @@
 #import "UAAnalytics.h"
 
 //total size in bytes that the event queue is allowed to grow to.
-#define kMaxTotalDBSizeBytes 5*1024*1024 // local max of 5MB
-#define kMinTotalDBSizeBytes 10*1024     // local min of 10KB
+#define kMaxTotalDBSizeBytes (NSInteger)5*1024*1024 // local max of 5MB
+#define kMinTotalDBSizeBytes (NSInteger)10*1024     // local min of 10KB
 
 // total size in bytes that a given event post is allowed to send.
-#define kMaxBatchSizeBytes 500*1024      // local max of 500KB
-#define kMinBatchSizeBytes 1024          // local min of 1KB
+#define kMaxBatchSizeBytes (NSInteger)500*1024      // local max of 500KB
+#define kMinBatchSizeBytes (NSInteger)1024          // local min of 1KB
 
 // maximum amount of time in seconds that events should queue for
-#define kMaxWaitSeconds 14*24*3600      // local max of 14 days
-#define kMinWaitSeconds 7*24*3600       // local min of 7 days
+#define kMaxWaitSeconds (NSInteger)14*24*3600      // local max of 14 days
+#define kMinWaitSeconds (NSInteger)7*24*3600       // local min of 7 days
 
 // The actual amount of time in seconds that elapse between event-server posts
-#define kMinBatchIntervalSeconds 60        // local min of 60s
-#define kMaxBatchIntervalSeconds 7*24*3600  // local max of 7 days
+#define kMinBatchIntervalSeconds (NSInteger)60        // local min of 60s
+#define kMaxBatchIntervalSeconds (NSInteger)7*24*3600  // local max of 7 days
 
 // minimum amount of time between background location events
 #define kMinBackgroundLocationIntervalSeconds 900 // 900 seconds = 15 minutes
@@ -67,7 +67,7 @@
 @property (nonatomic, assign) NSInteger maxBatchSize;
 @property (nonatomic, assign) NSInteger maxWait;
 @property (nonatomic, assign) NSInteger minBatchInterval;
-@property (nonatomic, assign) NSInteger databaseSize;
+@property (nonatomic, assign) NSUInteger databaseSize;
 @property (nonatomic, assign) NSTimeInterval oldestEventTime;
 @property (nonatomic, assign) UIBackgroundTaskIdentifier sendBackgroundTask;
 @property (nonatomic, retain) UAConfig *config;
