+++ Airship/Common/UAObservable.m
@@ -100,7 +100,7 @@
     }
 }
 
--(int)countObservers {
+-(NSUInteger)countObservers {
     return [self.observers count];
 }
